Matrix, Translation, Scale, and Rotation default changed to None
The chronos docs state that either matrix or TRS (Translation,
Rotation, and Scale) can be set but not both. It also says that
identity values are assumed if no content is entered. As such
setting None as the default value.

This fixes #27 on the main pygltflib repo. It doesn't seem to
affect any other functionality in the library. Is the librry starts
these values to transform objects then I suppose their needs to be
a "get_matrix" method that returns the matrix if defined, if not
returns T*R*S where T,R,S are either the matrix defined by their
value or the identity matrix if they are None.